Fixed scroll on category select
